requirements.txt
agency-swarm[fastapi]>=1.0.2
fastapi
uvicorn

# Document Processing
docling>=2.0.0
docling-core>=2.0.0

# Web Crawling
crawl4ai>=0.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Embeddings
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.1.0

# Database
asyncpg>=0.29.0
pgvector>=0.2.4

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.5.0
tiktoken>=0.5.0